requirements.txt
stripe
python-dotenv
pandas